Module layers::engine

Contains all the logic for running the animations, execute the scheduled changes to the models, and prepare the render tree The scene is drawn in 4 stages:

The layout step calculates the dimensions and position of a node and generates a transformation Matrix

The draw step generates a displaylist

The render step uses the displaylist to generate a texture of the node

The compose step generates the final image using the textures

Modules

animation

node

storage

Structs

AnimationRef

HandlerRef

LayersEngine

Main struct to interact with the engine

NodeRef

An identifier for a node in the three storage

PointerCallback

Timestamp

TransactionRef

Enums

TransactionEventType

Traits

Command

A trait for objects that can be exectuded by the engine.

SyncCommand