Module layers::engine::node

Structs

DrawCache

SceneNode is the main data structure for the engine. It contains a model that can be rendered, and a layout node that can be used to position and size the model. As well it contains the data structures that are used to cache the rendering of the model. Caching is done using skia Picture.

RenderableFlags

SceneNode

Traits

ContainsPoint

DrawCacheManagement

A trait for Nodes to expose their cache management