Struct layers::layers::BufferRequest

pub struct BufferRequest {
    pub screen_rect: Rect<usize>,
    pub page_rect: Rect<f32>,
    pub content_age: ContentAge,
    pub native_surface: Option<NativeSurface>,
}

A request from the compositor to the renderer for tiles that need to be (re)displayed.

Fields
screen_rect: Rect<usize>

The rect in pixels that will be drawn to the screen

page_rect: Rect<f32>

The rect in page coordinates that this tile represents

content_age: ContentAge

The content age of that this BufferRequest corresponds to.

native_surface: Option<NativeSurface>

A cached NativeSurface that can be used to avoid allocating a new one.

Methods

impl BufferRequest

fn new(screen_rect: Rect<usize>, page_rect: Rect<f32>, content_age: ContentAge) -> BufferRequest