Struct layers::layers::Layer

pub struct Layer<T> {
    pub children: RefCell<Vec<Rc<Layer<T>>>>,
    pub transform: RefCell<Matrix4<f32>>,
    pub tile_size: usize,
    pub extra_data: RefCell<T>,
    pub bounds: RefCell<TypedRect<LayerPixel, f32>>,
    pub content_age: RefCell<ContentAge>,
    pub content_offset: RefCell<TypedPoint2D<LayerPixel, f32>>,
    pub masks_to_bounds: RefCell<bool>,
    pub background_color: RefCell<Color>,
    // some fields omitted
}

Fields

children

transform

tile_size

extra_data

bounds

The boundaries of this layer in the coordinate system of the parent layer.

content_age

A monotonically increasing counter that keeps track of the current content age.

content_offset

The content offset for this layer in unscaled layer pixels.

masks_to_bounds

Whether this layer clips its children to its boundaries.

background_color

The background color for this layer.

Methods

impl<T> Layer<T>

fn new(bounds: TypedRect<LayerPixel, f32>, tile_size: usize, background_color: Color, data: T) -> Layer<T>

fn children<'a>(&'a self) -> RefMut<'a, Vec<Rc<Layer<T>>>>

fn add_child(&self, new_child: Rc<Layer<T>>)

fn get_buffer_requests(&self, rect_in_layer: TypedRect<LayerPixel, f32>, scale: ScaleFactor<LayerPixel, DevicePixel, f32>) -> Vec<BufferRequest>

fn resize(&self, new_size: TypedSize2D<LayerPixel, f32>)

fn add_buffer(&self, tile: Box<LayerBuffer>)

fn collect_unused_buffers(&self) -> Vec<Box<LayerBuffer>>

fn collect_buffers(&self) -> Vec<Box<LayerBuffer>>

fn contents_changed(&self)

fn create_textures(&self, graphics_context: &NativeCompositingGraphicsContext)

fn do_for_all_tiles<F: Fn(&Tile)>(&self, f: F)