Struct layers::layers::LayerBufferSet

pub struct LayerBufferSet {
    pub buffers: Vec<Box<LayerBuffer>>,
}

A set of layer buffers. This is an atomic unit used to switch between the front and back buffers.

Fields

buffers

Methods

impl LayerBufferSet

fn mark_will_leak(&mut self)

Notes all buffer surfaces will leak if not destroyed via a call to destroy.