Struct layers::platform::egl::surface::EGLImageNativeSurface

pub struct EGLImageNativeSurface {
    image: Option<EGLImageKHR>,
    bitmap: Option<Vec<u8>>,
    will_leak: bool,
    pub size: Size2D<i32>,
}

Fields

image

An EGLImage for the case of GPU rendering.

bitmap

A heap-allocated bitmap for the case of CPU rendering.

will_leak

Whether this pixmap will leak if the destructor runs. This is for debugging purposes.

size

The size of this surface.

Methods

impl EGLImageNativeSurface

fn new(_: &NativeDisplay, size: Size2D<i32>) -> EGLImageNativeSurface

fn bind_to_texture(&self, _: &NativeDisplay, texture: &Texture)

This may only be called on the compositor side.

fn upload(&mut self, _: &NativeDisplay, data: &[u8])

This may only be called on the painting side.

fn get_id(&self) -> isize

fn destroy(&mut self, _graphics_context: &NativeDisplay)

fn mark_will_leak(&mut self)

fn mark_wont_leak(&mut self)

fn gl_rasterization_context(&mut self, _gl_context: Arc<GLContext>) -> Option<GLRasterizationContext>