Enum layers::platform::linux::surface::NativeDisplay

pub enum NativeDisplay {
    EGL(EGLDisplayInfo),
    GLX(GLXDisplayInfo),
}

Variants
EGL(EGLDisplayInfo)GLX(GLXDisplayInfo)
Methods

impl NativeDisplay

fn new(display: *mut Display) -> NativeDisplay

fn compositor_visual_info(display: *mut Display) -> (*mut XVisualInfo, Option<GLXFBConfig>)

Chooses the compositor visual info using the same algorithm that the compositor uses.

FIXME(pcwalton): It would be more robust to actually have the compositor pass the visual.

fn get_compatible_configuration(display: *mut Display, configs: *mut GLXFBConfig, number_of_configs: i32) -> (*mut XVisualInfo, Option<GLXFBConfig>)

fn need_to_find_32_bit_depth_visual(display: *mut Display) -> bool

fn platform_display_data(&self) -> PlatformDisplayData

fn new_egl_display() -> NativeDisplay