Struct layers::platform::linux::surface::GLXDisplayInfo

pub struct GLXDisplayInfo {
    pub display: *mut Display,
    visual_info: *mut XVisualInfo,
    framebuffer_configuration: Option<GLXFBConfig>,
}

The display, visual info, and framebuffer configuration. This is needed in order to bind to a texture on the compositor side. This holds only a weak reference to the display and does not close it.

FIXME(pcwalton): Unchecked weak references are bad and can violate memory safety. This is hard to fix because the Display is given to us by the native windowing system, but we should fix it someday. FIXME(pcwalton): Mark nonsendable.

Fields

display

visual_info

framebuffer_configuration