Struct layers::platform::linux::surface::NativeCompositingGraphicsContext

pub struct NativeCompositingGraphicsContext {
    // some fields omitted
}

The display, visual info, and framebuffer configuration. This is needed in order to bind to a texture on the compositor side. This holds only a weak reference to the display and does not close it.

FIXME(pcwalton): Unchecked weak references are bad and can violate memory safety. This is hard to fix because the Display is given to us by the native windowing system, but we should fix it someday.

FIXME(pcwalton): Mark nonsendable.

Methods

impl NativeCompositingGraphicsContext

fn from_display(display: *mut Display) -> NativeCompositingGraphicsContext

Creates a native graphics context from the given X display connection. This uses GLX. Only the compositor is allowed to call this.