Struct layers::platform::linux::surface::NativePaintingGraphicsContext

pub struct NativePaintingGraphicsContext {
    pub display: *mut Display,
    // some fields omitted
}

The display and visual info. This is needed in order to upload on the painting side. This holds a weak reference to the display and will not close it when done.

FIXME(pcwalton): Mark nonsendable.

Fields

display

Methods

impl NativePaintingGraphicsContext

fn from_metadata(metadata: &NativeGraphicsMetadata) -> NativePaintingGraphicsContext