Enum layers::platform::surface::NativeSurface

pub enum NativeSurface {
    MemoryBuffer(MemoryBufferNativeSurface),
    Pixmap(PixmapNativeSurface),
}

Variants

MemoryBuffer

Pixmap

Methods

impl NativeSurface

fn new(native_context: &NativePaintingGraphicsContext, size: Size2D<i32>, stride: i32) -> NativeSurface

Creates a new native surface with uninitialized data.

impl NativeSurface

fn from_draw_target_backing(backing: DrawTargetBacking) -> NativeSurface

fn bind_to_texture(&self, native_context: &NativeCompositingGraphicsContext, texture: &Texture, size: Size2D<isize>)

Binds the surface to a GPU texture. Compositing task only.

fn upload(&mut self, native_context: &NativePaintingGraphicsContext, data: &[u8])

Uploads pixel data to the surface. Painting task only.

fn get_id(&self) -> isize

Returns an opaque ID identifying the surface for debugging.

fn destroy(&mut self, graphics_context: &NativePaintingGraphicsContext)

Destroys the surface. After this, it is an error to use the surface. Painting task only.

fn mark_will_leak(&mut self)

Records that the surface will leak if destroyed. This is done by the compositor immediately after receiving the surface.

fn mark_wont_leak(&mut self)

Marks the surface as not leaking. The painting task and the compositing task call this when they are certain that the surface will not leak. For example:

When sending buffers back to the render task, either the render task will receive them or the render task has crashed. In the former case, they're the render task's responsibility, so this is OK. In the latter case, the kernel or window server is going to clean up the layer buffers. Either way, no leaks.

If the compositor is shutting down, the render task is also shutting down. In that case it will destroy all its pixmaps, so no leak.

If the painting task is sending buffers to the compositor, then they are marked as not leaking, because of the possibility that the compositor will die before the buffers are destroyed.

This helps debug leaks. For performance this may want to become a no-op in the future.