Struct layers::platform::surface::MemoryBufferNativeSurface

pub struct MemoryBufferNativeSurface {
    // some fields omitted
}

Methods

impl MemoryBufferNativeSurface

fn new(_: &NativePaintingGraphicsContext, _: Size2D<i32>, _: i32) -> MemoryBufferNativeSurface

fn bind_to_texture(&self, _: &NativeCompositingGraphicsContext, texture: &Texture, size: Size2D<isize>)

This may only be called on the compositor side.

fn upload(&mut self, _: &NativePaintingGraphicsContext, data: &[u8])

This may only be called on the painting side.

fn get_id(&self) -> isize

fn destroy(&mut self, _: &NativePaintingGraphicsContext)

fn mark_will_leak(&mut self)

fn mark_wont_leak(&mut self)