Enum layers::prelude::taffy::TrackSizingFunction

pub enum TrackSizingFunction {
    Single(MinMax<MinTrackSizingFunction, MaxTrackSizingFunction>),
    Repeat(GridTrackRepetition, Vec<MinMax<MinTrackSizingFunction, MaxTrackSizingFunction>>),
}

The sizing function for a grid track (row/column) See https://developer.mozilla.org/en-US/docs/Web/CSS/grid-template-columns

Variants

Single(MinMax<MinTrackSizingFunction, MaxTrackSizingFunction>)

A single non-repeated track

Repeat(GridTrackRepetition, Vec<MinMax<MinTrackSizingFunction, MaxTrackSizingFunction>>)

Automatically generate grid tracks to fit the available space using the specified definite track lengths Only valid if every track in template (not just the repetition) has a fixed size.

Implementations

impl TrackSizingFunction

pub fn is_auto_repetition(&self) -> bool

Whether the track definition is a auto-repeated fragment