Function layers::prelude::taffy::compute_cached_layout

pub fn compute_cached_layout<Tree, ComputeFunction>(
    tree: &mut Tree,
    node: NodeId,
    inputs: LayoutInput,
    compute_uncached: ComputeFunction,
) -> LayoutOutput
where
    Tree: LayoutPartialTree + ?Sized,
    ComputeFunction: FnMut(&mut Tree, NodeId, LayoutInput) -> LayoutOutput,

Attempts to find a cached layout for the specified node and layout inputs.

Uses the provided closure to compute the layout (and then stores the result in the cache) if no cached layout is found.