Struct layers::scene::Scene

pub struct Scene<T> {
    pub root: Option<Rc<Layer<T>>>,
    pub viewport: TypedRect<DevicePixel, f32>,
    pub scale: ScaleFactor<LayerPixel, DevicePixel, f32>,
}

Fields
root: Option<Rc<Layer<T>>> viewport: TypedRect<DevicePixel, f32> scale: ScaleFactor<LayerPixel, DevicePixel, f32>

The scene scale, to allow for zooming and high-resolution painting.

Methods

impl<T> Scene<T>

fn new(viewport: TypedRect<DevicePixel, f32>) -> Scene<T>

fn get_buffer_requests_for_layer(&mut self, layer: Rc<Layer<T>>, dirty_rect: TypedRect<LayerPixel, f32>, viewport_rect: TypedRect<LayerPixel, f32>, layers_and_requests: &mut Vec<(Rc<Layer<T>>, Vec<BufferRequest>)>, unused_buffers: &mut Vec<Box<LayerBuffer>>)

fn get_buffer_requests(&mut self, requests: &mut Vec<(Rc<Layer<T>>, Vec<BufferRequest>)>, unused_buffers: &mut Vec<Box<LayerBuffer>>)

fn mark_layer_contents_as_changed_recursively_for_layer(&self, layer: Rc<Layer<T>>)

fn mark_layer_contents_as_changed_recursively(&self)

fn set_root_layer_size(&self, new_size: TypedSize2D<DevicePixel, f32>)

fn get_memory_usage(&self) -> usize

Calculate the amount of memory used by all the layers in the scene graph. The memory may be allocated on the heap or in GPU memory.