Struct layers::tiling::Tile

pub struct Tile {
    buffer: Option<Box<LayerBuffer>>,
    content_age_of_pending_buffer: Option<ContentAge>,
    pub texture: Texture,
    pub bounds: Option<TypedRect<LayerPixel, f32>>,
}

Fields

buffer

The buffer displayed by this tile.

content_age_of_pending_buffer

The content age of any pending buffer request to avoid re-requesting a buffer while waiting for it to come back from rendering.

texture

A handle to the GPU texture.

bounds

The tile boundaries in the parent layer coordinates.

Methods

impl Tile

fn new() -> Tile

fn should_use_new_buffer(&self, new_buffer: &Box<LayerBuffer>) -> bool

fn replace_buffer(&mut self, buffer: Box<LayerBuffer>) -> Option<Box<LayerBuffer>>

fn create_texture(&mut self, display: &NativeDisplay)

fn should_request_buffer(&self, content_age: ContentAge) -> bool