Struct layers::tiling::TileGrid

pub struct TileGrid {
    pub tiles: HashMap<Point2D<usize>, Tile>,
    tile_size: Length<DevicePixel, usize>,
    unused_buffers: Vec<Box<LayerBuffer>>,
}

Fields
tiles: HashMap<Point2D<usize>, Tile> tile_size: Length<DevicePixel, usize>

The size of tiles in this grid in device pixels.

unused_buffers: Vec<Box<LayerBuffer>>
Methods

impl TileGrid

fn new(tile_size: usize) -> TileGrid

fn get_rect_for_tile_index(&self, tile_index: Point2D<usize>, current_layer_size: TypedSize2D<DevicePixel, f32>) -> TypedRect<DevicePixel, usize>

fn take_unused_buffers(&mut self) -> Vec<Box<LayerBuffer>>

fn add_unused_buffer(&mut self, buffer: Option<Box<LayerBuffer>>)

fn tile_intersects_rect(&self, tile_index: &Point2D<usize>, test_rect: &Rect<f32>, current_layer_size: TypedSize2D<DevicePixel, f32>, layer_world_origin: &Point2D<f32>, layer_transform: &Matrix4D<f32>) -> bool

fn mark_tiles_outside_of_rect_as_unused(&mut self, rect: TypedRect<DevicePixel, f32>, layer_world_origin: &Point2D<f32>, layer_transform: &Matrix4D<f32>, current_layer_size: TypedSize2D<DevicePixel, f32>)

fn get_buffer_request_for_tile(&mut self, tile_index: Point2D<usize>, current_layer_size: TypedSize2D<DevicePixel, f32>, current_content_age: ContentAge) -> Option<BufferRequest>

fn get_buffer_requests_in_rect(&mut self, dirty_rect: TypedRect<DevicePixel, f32>, viewport: TypedRect<DevicePixel, f32>, current_layer_size: TypedSize2D<DevicePixel, f32>, layer_world_origin: &Point2D<f32>, layer_transform: &Matrix4D<f32>, current_content_age: ContentAge) -> Vec<BufferRequest>

Returns buffer requests inside the given dirty rect, and simultaneously throws out tiles outside the given viewport rect.

fn get_tile_index_for_point(&self, point: Point2D<usize>) -> Point2D<usize>

fn add_buffer(&mut self, buffer: Box<LayerBuffer>)

fn do_for_all_tiles<F>(&self, f: F) where F: FnMut(&Tile)

fn collect_buffers(&mut self) -> Vec<Box<LayerBuffer>>

fn create_textures(&mut self, display: &NativeDisplay)

fn get_memory_usage(&self) -> usize

Calculate the amount of memory used by all the tiles in the tile grid. The memory may be allocated on the heap or in GPU memory.