Struct layers::layers::Layer

pub struct Layer<T> {
    pub children: RefCell<Vec<Rc<Layer<T>>>>,
    pub transform: RefCell<Matrix4D<f32>>,
    pub perspective: RefCell<Matrix4D<f32>>,
    pub tile_size: usize,
    pub extra_data: RefCell<T>,
    tile_grid: RefCell<TileGrid>,
    pub bounds: RefCell<TypedRect<LayerPixel, f32>>,
    pub content_age: RefCell<ContentAge>,
    pub content_offset: RefCell<TypedPoint2D<LayerPixel, f32>>,
    pub masks_to_bounds: RefCell<bool>,
    pub background_color: RefCell<Color>,
    pub opacity: RefCell<f32>,
    pub establishes_3d_context: bool,
    pub transform_state: RefCell<TransformState>,
}

Fields
children: RefCell<Vec<Rc<Layer<T>>>> transform: RefCell<Matrix4D<f32>> perspective: RefCell<Matrix4D<f32>> tile_size: usize extra_data: RefCell<T> tile_grid: RefCell<TileGrid> bounds: RefCell<TypedRect<LayerPixel, f32>>

The boundaries of this layer in the coordinate system of the parent layer.

content_age: RefCell<ContentAge>

A monotonically increasing counter that keeps track of the current content age.

content_offset: RefCell<TypedPoint2D<LayerPixel, f32>>

The content offset for this layer in unscaled layer pixels.

masks_to_bounds: RefCell<bool>

Whether this layer clips its children to its boundaries.

background_color: RefCell<Color>

The background color for this layer.

opacity: RefCell<f32>

The opacity of this layer, from 0.0 (fully transparent) to 1.0 (fully opaque).

establishes_3d_context: bool

Whether this stacking context creates a new 3d rendering context.

transform_state: RefCell<TransformState>

Collection of state related to transforms for this layer.

Methods

impl<T> Layer<T>

fn new(bounds: TypedRect<LayerPixel, f32>, tile_size: usize, background_color: Color, opacity: f32, establishes_3d_context: bool, data: T) -> Layer<T>

fn children(&self) -> RefMut<Vec<Rc<Layer<T>>>>

fn add_child(&self, new_child: Rc<Layer<T>>)

fn remove_child_at_index(&self, index: usize)

fn get_buffer_requests(&self, rect_in_layer: TypedRect<LayerPixel, f32>, viewport_in_layer: TypedRect<LayerPixel, f32>, scale: ScaleFactor<LayerPixel, DevicePixel, f32>) -> Vec<BufferRequest>

Returns buffer requests inside the given dirty rect, and simultaneously throws out tiles outside the given viewport rect.

fn resize(&self, new_size: TypedSize2D<LayerPixel, f32>)

fn add_buffer(&self, tile: Box<LayerBuffer>)

fn collect_unused_buffers(&self) -> Vec<Box<LayerBuffer>>

fn collect_buffers(&self) -> Vec<Box<LayerBuffer>>

fn contents_changed(&self)

fn create_textures(&self, display: &NativeDisplay)

fn do_for_all_tiles<F: FnMut(&Tile)>(&self, f: F)

fn update_transform_state(&self, parent_transform: &Matrix4D<f32>, parent_perspective: &Matrix4D<f32>, parent_origin: &Point2D<f32>)

fn get_memory_usage(&self) -> usize

Calculate the amount of memory used by this layer and all its children. The memory may be allocated on the heap or in GPU memory.