Struct layers::layers::LayerBuffer

pub struct LayerBuffer {
    pub native_surface: NativeSurface,
    pub rect: Rect<f32>,
    pub screen_pos: Rect<usize>,
    pub resolution: f32,
    pub painted_with_cpu: bool,
    pub content_age: ContentAge,
}

Fields
native_surface: NativeSurface

The native surface which can be shared between threads or processes. On Mac this is an IOSurface; on Linux this is an X Pixmap; on Android this is an EGLImageKHR.

rect: Rect<f32>

The rect in the containing RenderLayer that this represents.

screen_pos: Rect<usize>

The rect in pixels that will be drawn to the screen.

resolution: f32

The scale at which this tile is rendered

painted_with_cpu: bool

Whether or not this buffer was painted with the CPU rasterization.

content_age: ContentAge

The content age of that this buffer request corresponds to.

Methods

impl LayerBuffer

fn get_mem(&self) -> usize

Returns the amount of memory used by the tile

fn is_valid(&self, scale: f32) -> bool

Returns true if the tile is displayable at the given scale

fn get_size_2d(&self) -> Size2D<usize>

Returns the Size2D of the tile

fn mark_wont_leak(&mut self)

Marks the layer buffer as not leaking. See comments on NativeSurfaceMethods::mark_wont_leak for how this is used.

fn destroy(self, display: &NativeDisplay)

Destroys the layer buffer. Painting task only.